login & testing interface QoL changes
updated login and use sessionStorage to keep track of the person logging
in. This will be used in the email sent out. Email buttons are all
questions (email code functionality soon to come). This will email out
whatever has been answered thus far. You can choose which question to
start on so you do not have to traverse the questions to get to a
certain question (this is for students that are resuming a test).